\name{rice_normal}
\alias{rice_normal}
\docType{data}
\title{
The gene expression data of rice under control  or normal condition
}
\description{
rice_normal data has the gene expression values of 200 genes over 20 samples. The row names of the data are the names of the probes.
}
\usage{data(rice_normal)}
\format{
  A data frame with 200 rows and 20 columns.
  }
\details{rice_normal data has the gene expression values of 200 genes identified by Wang et al. (2013). After loading the data, the dataset is available in R in the name of rice_normal.
}
\source{Wang J, Chen L, Wang Y, Zhang J, Liang Y, Xu D (2013) A Computational systems biology study for understanding salt tolerance mechanism in Rice. PLoS one 8(6): e64929. doi:10.1371/journal.pone.0064929.
}
\examples{
data(rice_normal)
}
\keyword{datasets}
